Refactor common base class for object permission commands
The permission adding and removing commands are quite similar.
